+++ wiki/wiki_md/about/changelog.md
@@ -18,6 +18,7 @@ This page documents the changes and fixes made in Quantum Things, compared to th
 
 - Fixed Biome Sensor not working when held in off-hand.
 - Fixed Nature Core, Bonemealing and Animal Spawning not being centered on the Nature Core (introduced in 1.0.0).
+- Fixed plate item entities being way too large when dropped as items.
 
 ## v1.0.1