+++ wiki/wiki_md/about/changelog.md
@@ -15,6 +15,9 @@ This page documents the changes and fixes made in Quantum Things, compared to th
 - Added the ability to configure the Spectre Energy Injector capacity, Spectre Coil/Charger transfer rates, and whether the Genesis Spectre Coil generates energy or transfers it.
 - Made the ID Card crafting recipe shapeless.
 - Re-added Spectre Armor (WIP).
+- Re-added Biome Painter and Biome Capsule from 1.7.10.
+- Re-added Obsidian Stick from 1.7.10, now used for Spectre Weapons and the Biome Painter.
+- Added the Spectre Hoe.
 
 ### Fixes