+++ wiki/wiki_md/about/changelog.md
@@ -16,7 +16,7 @@ This page documents the changes and fixes made in Quantum Things, compared to th
 - Made the ID Card crafting recipe shapeless.
 - Re-added Spectre Armor (WIP).
 - Re-added Biome Painter and Biome Capsule from 1.7.10.
-- Re-added Obsidian Stick from 1.7.10, now used for Spectre Weapons and the Biome Painter.
+- Re-added Obsidian Stick from 1.7.10, now used for Spectre Tools and the Biome Painter.
 - Added the Spectre Hoe.
 
 ### Fixes